Fix remaining int64_t conversions and add FFmpeg profile fallbacks
Linux - MKV and AVI offset conversions:
- element.offset and chunk.offset (int64_t) to QVariant ambiguous
- Added explicit cast: static_cast<qlonglong>(offset)
- Same issue as MP4 atoms, now fixed for all container formats

Windows - FFmpeg profile constants missing:
- BtbN FFmpeg builds may not define FF_PROFILE_* constants
- Added fallback #define with ITU-T standard values:
  - H.264: Baseline(66), Main(77), High(100), High10(110), etc.
  - H.265: Main(1), Main10(2), MainStillPicture(3)
- Constants match ITU-T H.264/H.265 specifications
- Only defined if not already present (#ifndef guards)

This ensures:
- ✅ Builds with any FFmpeg version (with or without profile constants)
- ✅ Profile validation works correctly with standard values
- ✅ All int64_t to QVariant conversions explicit

+++ src/panels/explorerpanel.cpp
@@ -306,8 +306,8 @@ void ExplorerPanel::addMKVElementNodes(QTreeWidgetItem* parent, const QVector<EB
         elementItem->setText(2, QString::number(element.percentage, 'f', 2));
         elementItem->setCheckState(0, Qt::Checked);
 
-        // Store element offset as user data
-        elementItem->setData(0, Qt::UserRole, element.offset);
+        // Store element offset as user data (cast to qlonglong for Qt 6.5.3)
+        elementItem->setData(0, Qt::UserRole, static_cast<qlonglong>(element.offset));
 
         // Recursively add children
         if (!element.children.isEmpty()) {
@@ -354,8 +354,8 @@ void ExplorerPanel::addAVIChunkNodes(QTreeWidgetItem* parent, const QVector<AVIC
         chunkItem->setText(2, QString::number(chunk.percentage, 'f', 2));
         chunkItem->setCheckState(0, Qt::Checked);
 
-        // Store chunk offset as user data
-        chunkItem->setData(0, Qt::UserRole, chunk.offset);
+        // Store chunk offset as user data (cast to qlonglong for Qt 6.5.3)
+        chunkItem->setData(0, Qt::UserRole, static_cast<qlonglong>(chunk.offset));
 
         // Recursively add children
         if (!chunk.children.isEmpty()) {
